examples/tracker/pyproject.toml
[project]
name = "dora-cotracker"
version = "0.1.0"
description = "Dora node for point tracking using Facebook's CoTracker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dora-rs>=0.3.6",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "pyarrow>=8.0.0",
    "torch>=2.0.0",
    "cotracker @ git+https://github.com/facebookresearch/co-tracker.git",
]

[project.scripts]
dora-cotracker = "dora_cotracker.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
